% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cameratraps_path_constructor.R
\name{construct_path_to_site_folder_from_site_code}
\alias{construct_path_to_site_folder_from_site_code}
\title{Get Path to Site Folder from Site Code}
\usage{
construct_path_to_site_folder_from_site_code(sitecode_df, path)
}
\arguments{
\item{sitecode_df}{A data frame that contains a column named \code{site_code} and
(optionally) a column named \code{collection_folder}. This data frame is usually
generated by the function \code{extract_site_code_from_collection_folder()}.}

\item{path}{A string that is a valid file path to the location of the
top-level folder of the 'cameratraps' or 'cameratraps2' folders.}
}
\value{
Returns a data frame with a new column \code{site_folder} which contains a
path to the site folder.
}
\description{
\code{construct_path_to_site_folder_from_site_code()} returns a data frame with the path
to the site folder in a new column named \code{site_folder} when given a data
frame that has a column named \code{site_code} and a column named \code{path}.
}
\examples{
library(GrazingCameratraps)

df <- data.frame(
  site_code = "A51",
  site_folder = "G:/cameratraps/fiftyone/timelapse"
  )
construct_path_to_site_folder_from_site_code(df, path = "G:")

}
