% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cameratraps_path_constructor.R
\name{site_folder_from_sitecode}
\alias{site_folder_from_sitecode}
\title{site_folder_from_sitecode}
\usage{
site_folder_from_sitecode(sitecode_df, path)
}
\arguments{
\item{path}{}
}
\description{
site_folder_from_sitecode
}
