% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject_text_files_to_cameratraps_path.R
\name{subject_text_files_to_cameratraps_path}
\alias{subject_text_files_to_cameratraps_path}
\title{subject_text_files_to_cameratraps_path}
\usage{
subject_text_files_to_cameratraps_path(
  subject_txt_df,
  cameratraps_folders_to_match_df
)
}
\arguments{
\item{subject_txt_df}{}
}
\value{
Returns a data frame with the
}
\description{
A short description...
}
