% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cameratraps_path_constructor.R
\name{construct_path_from_collection_and_site_folders}
\alias{construct_path_from_collection_and_site_folders}
\title{construct_path_from_collection_and_site_folders}
\usage{
construct_path_from_collection_and_site_folders(sitecode_df)
}
\arguments{
\item{sitecode_df}{}
}
\description{
\code{construct_path_from_collection_and_site_folders} constructs a file path by combining the path to the site_folder and appending the collection folder onto the path.
}
